+++ blocks/b-productswithselectors/b-productswithselectors.js
@@ -16,7 +16,8 @@
  * - https://www.bitdefender.com/media/html/business/RansomwareTrial/new.html
  */
 
-import { updateProductsList, productAliases } from '../../scripts/scripts.js';
+import { productAliases } from '../../scripts/scripts.js';
+import { updateProductsList } from '../../scripts/utils.js';
 
 export default function decorate(block) {
   /// ///////////////////////////////////////////////////////////////////////
