+++ scripts/delayed.js
@@ -2,26 +2,27 @@
 import { sampleRUM } from './lib-franklin.js';
 
 import { sendAnalyticsPageLoadedEvent } from './adobeDataLayer.js';
-import { addScript, instance } from './utils.js';
+import {
+  addScript, instance, isZuoraNL, productsList, showPrices,
+} from './utils.js';
 import initZuoraNL from './zuora.js';
 
 // Core Web Vitals RUM collection
 sampleRUM('cwv');
 
 // add more delayed functionality here
-const isZuoraNL = window.DEFAULT_LANGUAGE === 'nl';
-
-if (isZuoraNL) {
+// todo: logic for zuora
+if (isZuoraNL()) {
   // for NL - Zuora
   window.config = initZuoraNL.config();
-  addScript('https://checkout.bitdefender.com/static/js/sdk.js', {}, 'defer', () => {
+  addScript('https://checkout.bitdefender.com/static/js/sdk.js', {}, 'async', () => {
     console.log('initZuoraNL');
-    // if (productsList.length) {
-    //   productsList.forEach(async (item) => {
-    //     const zuoraResult = await initZuoraNL.loadProduct(item);
-    //     showPrices(zuoraResult);
-    //   });
-    // }
+    if (productsList.length) {
+      productsList.forEach(async (item) => {
+        const zuoraResult = await initZuoraNL.loadProduct(item);
+        showPrices(zuoraResult);
+      });
+    }
   });
 }
